refactor: move drawer-content and drawer-side to the child layout
